fixed cache path issue and fixed str casting bug
cache path was throwing an error if you imported the library from
somewhere else as it was always trying to cache data in the current
directory.

the param num_regions was a string and balltree code does not
automatically cast this to an int in a comparison for number
of neighbours requested. thus, an error occurs.