added using api.get_geojson()
this change was a bit hacky. it requires going through the shapely
library to get have GeoDataFrame properly read in the geojson, as
GeoDataFrame has no support for loading from a Dict. This opens a
second issue of some districts being composed of MultiPolygons which
end up appearing as multiple rows in the GeoDataFrame. not sure how to
merge these into one "row", or if the trick is just to assign them all
the same district ID and colour them in accordingly each time.