handle case where data cannot be normalized return absolute weights
Avoids a division by zero failure when the total is 0.

We do not enforce any assumption on the given series, but in basic
cases where the weighting series is non-negative, total=0 happens when
the data is all 0 or missing so returning 0 weights is most convenient
in applications where there are multiple sets of weights.