use sphinx-versions to avoid unicode errors
We use sphinxcontrib-versioning to build the documentation on
developers.gro-intelligence.com. It hasn't been updated since 2016 and
doesn't work on latest versions of Python 3 we use in Shippable. So,
we're using Python 2 to build our documentation. Sometimes, this flakily
results in unexpected unicode-related errors (e.g.
https://app.shippable.com/github/gro-intelligence/api-client/runs/3833/1/console),
which blocks our ability to merge PRs.

This change switches to using sphinx-versions instead, a fork of
sphinxcontrib-versioning which has at least been updated since 2019 and
has some unicode-related bugfixes:
https://github.com/Smile-SA/sphinx-versions