fix doc gen UnicodeDecodeError on Shippable
sphinxcontrib-versioning claims to only support up to Python 3.5, but
manual testing on Python 3.7 shows that it works. (It hasn't been
updated since 2016, so that's probably why it doesn't list support for
recent versions.)

This should fix the nondeterministic-seeming documentation generation
errors we see on Shippable, which has been blocking PR merges.
CLEWS-32963