disabling Shippable on gh-pages step 1
Shippable keeps running (and failing) on the gh-pages branch
unnecessarily. We need to add a shippable.yml file to the gh-pages
branch itself to disable Shippable. But first, we need to keep our docs
automation to not remove the file automatically.

Also removes the (useless) `branches: except: gh-pages` config from the
main development branch's shippable.yml.