feat: add test_notebooks.sh script
This script uses the nbval pytest plugin to check that (some of) our
notebooks execute without errors. A further subset of our notebooks have
strict checking that re-executing the code in the notebook resuts in the
same outputs as previously saved.

Note: this is not currently automated, as it requires a working
GROAPI_TOKEN value and executing all the notebooks (using a separate
venv for each) takes non-trivial run time.