fix: each GroClient has its own AsyncHTTPClient
This should finally, properly fix the `fetch() called on closed
AsyncHTTPClient` errors that we've been seeing since #257. The root
problem is that AsyncHTTPClient has an internal instance cache, so
creating a new GroClient when a previous one exists would reuse the
internal AsyncHTTPClient. If the previous GroClient falls out of scope,
the (shared) AsyncHTTPClient is closed (via GroClient.__del__),
resulting in errors when trying to use the new GroClient.

- See the Tornado docs for `AsyncHTTPClient`, `force_instance`, and the
  `close` method:
  https://www.tornadoweb.org/en/stable/httpclient.html#tornado.httpclient.AsyncHTTPClient
- We explicitly started closing the AsyncHTTPClient in this PR (due to
  file descriptor leaking):
  https://github.com/gro-intelligence/api-client/pull/257
- This previous workaround fixed some instances of the error on Celery:
  https://github.com/gro-intelligence/api-client/pull/262
  It's still mysterious why that PR fixed things, and why more things on
  Celery weren't failing... Perhaps Celery does other things with
  IOLoops that result in this bug not being triggered sometimes.