feat: make GroClient() args optional
Changes GroClient.__init__ so that both API host and access token are
optional arguments. The API host defaults to the production API server
and the access token defaults to `$GROAPI_TOKEN`. This saves clients
(and us) from needing to spell out 'api.gro-intelligence.com' over and
over.

Also added docstring for the constructor, and added to developer docs.