add destructor checks for clean shutdown
This changes ensures _async_http_client and _ioloop are always defined,
and adds checks that they are set before trying to close them in the
destructor. Without this, if the constructor fails early (because
access_token isn't set), the destructor raises exceptions trying to
access and call methods on the unset values.

Noticed this running pytest on Python 3.9. For whatever reason, pytest
on Python 3.6 didn't complain.